requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
pypdf>=4.0.0
numpy>=1.26.0
pandas>=2.2.0